requirements.txt
flask
requests
feedparser
ics
python-dateutil
